Use 0-based id for default C selectors.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@6723 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ preprocessor/pre-post/ecs_pre_med.c
@@ -557,7 +557,7 @@ ecs_loc_pre_med__lit_maille(ecs_maillage_t   *maillage,
       ecs_warn();
       printf(_("The MED mesh contains %d elements of type %s\n"
                "which are ignored by the Preprocessor.\n\n"),
-             nbr_ele_med, nom_typ_ignore[ityp]);
+             (int)nbr_ele_med, nom_typ_ignore[ityp]);
 
     }
   }
